+++ src/main/java/hunternif/mc/atlas/core/AtlasData.java
@@ -1,5 +1,6 @@
 package hunternif.mc.atlas.core;
 
+import hunternif.mc.atlas.AntiqueAtlasMod;
 import hunternif.mc.atlas.network.PacketDispatcher;
 import hunternif.mc.atlas.network.client.MapDataPacket;
 import hunternif.mc.atlas.network.server.BrowsingPositionPacket;
@@ -9,7 +10,9 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
+import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.WorldSavedData;
+import net.minecraft.world.chunk.Chunk;
 import net.minecraftforge.common.util.Constants;
 
 import java.util.*;
@@ -27,7 +30,13 @@ public class AtlasData extends WorldSavedData {
 	private static final String TAG_BROWSING_X = "qBrowseX";
 	private static final String TAG_BROWSING_Y = "qBrowseY";
 	private static final String TAG_BROWSING_ZOOM = "qBrowseZoom";
-	
+
+	/** Maps dimension ID to biomeAnalyzer. */
+	private final Map<Integer, IBiomeDetector> biomeAnalyzers = new HashMap<>();
+	private final BiomeDetectorBase biomeDetectorOverworld = new BiomeDetectorBase();
+	private final BiomeDetectorNether biomeDetectorNether = new BiomeDetectorNether();
+	private final BiomeDetectorEnd biomeDetectorEnd = new BiomeDetectorEnd();
+
 	/** This map contains, for each dimension, a map of chunks the player
 	 * has seen. This map is thread-safe.
 	 * CAREFUL! Don't modify chunk coordinates that are already put in the map! */
@@ -39,8 +48,13 @@ public class AtlasData extends WorldSavedData {
 	
 	private NBTTagCompound nbt;
 
-	public AtlasData(String key) {
+	AtlasData(String key) {
 		super(key);
+
+		biomeDetectorOverworld.setScanPonds(AntiqueAtlasMod.settings.doScanPonds);
+		setBiomeDetectorForDimension(0, biomeDetectorOverworld);
+		setBiomeDetectorForDimension(-1, biomeDetectorNether);
+		setBiomeDetectorForDimension(1, biomeDetectorEnd);
 	}
 
 	@Override
@@ -91,6 +105,88 @@ public class AtlasData extends WorldSavedData {
 		compound.setTag(TAG_DIMENSION_MAP_LIST, dimensionMapList);
 		
 		return compound;
+	}
+
+	public void setBiomeDetectorForDimension(int dimension, IBiomeDetector biomeAnalyzer) {
+		biomeAnalyzers.put(dimension, biomeAnalyzer);
+	}
+
+	/** If not found, returns the analyzer for overworld. */
+	private IBiomeDetector getBiomeDetectorForDimension(int dimension) {
+		IBiomeDetector biomeAnalyzer = biomeAnalyzers.get(dimension);
+		return biomeAnalyzer == null ? biomeDetectorOverworld : biomeAnalyzer;
+	}
+
+	public void updateMapAroundPlayer(EntityPlayer player) {
+		// Update the actual map only so often:
+		int newScanInterval = Math.round(AntiqueAtlasMod.settings.newScanInterval * 20);
+		int rescanInterval = newScanInterval * AntiqueAtlasMod.settings.rescanRate;
+		if (player.ticksExisted % newScanInterval != 0) {
+			return;
+		}
+
+		int playerX = MathHelper.floor(player.posX) >> 4;
+		int playerZ = MathHelper.floor(player.posZ) >> 4;
+		ITileStorage seenChunks = this.getDimensionData(player.dimension);
+		IBiomeDetector biomeDetector = getBiomeDetectorForDimension(player.dimension);
+		int scanRadius = AntiqueAtlasMod.settings.scanRadius;
+
+		// Look at chunks around in a circular area:
+		for (double dx = -scanRadius; dx <= scanRadius; dx++) {
+			for (double dz = -scanRadius; dz <= scanRadius; dz++) {
+				if (dx*dx + dz*dz > scanRadius*scanRadius) {
+					continue; // Outside the circle
+				}
+				int x = (int)(playerX + dx);
+				int z = (int)(playerZ + dz);
+				Tile oldTile = seenChunks.getTile(x, z);
+
+				// Check if there's a custom tile at the location:
+				int biomeId = AntiqueAtlasMod.extBiomeData.getData().getBiomeIdAt(player.dimension, x, z);
+				// Custom tiles overwrite even the chunks already seen.
+
+				// If there's no custom tile, check the actual chunk:
+				if (biomeId == -1) {
+					Chunk chunk = player.getEntityWorld().getChunkFromChunkCoords(x, z);
+					// Force loading of chunk, if required:
+					if (AntiqueAtlasMod.settings.forceChunkLoading && !chunk.isLoaded()) {
+						player.getEntityWorld().getChunkProvider().provideChunk(x << 4, z << 4);
+					}
+					// Skip chunk if it hasn't loaded yet:
+					if (!chunk.isLoaded()) {
+						continue;
+					}
+
+					if (oldTile != null) {
+						// If the chunk has been scanned previously, only re-scan it so often:
+						if (!AntiqueAtlasMod.settings.doRescan || player.ticksExisted % rescanInterval != 0) {
+							continue;
+						}
+						biomeId = biomeDetector.getBiomeID(chunk);
+						if (biomeId == IBiomeDetector.NOT_FOUND) {
+							// If the new tile is empty, remove the old one:
+							this.removeTile(player.dimension, x, z);
+						} else if (oldTile.biomeID != biomeId) {
+							// Only update if the old tile's biome ID doesn't match the new one:
+							this.setTile(player.dimension, x, z, new Tile(biomeId));
+						}
+					} else {
+						// Scanning new chunk:
+						biomeId = biomeDetector.getBiomeID(chunk);
+						if (biomeId != IBiomeDetector.NOT_FOUND) {
+							this.setTile(player.dimension, x, z, new Tile(biomeId));
+						}
+					}
+				} else {
+					// Only update the custom tile if it doesn't rewrite itself:
+					if (oldTile == null || oldTile.biomeID != biomeId) {
+						this.setTile(player.dimension, x, z, new Tile(biomeId));
+						this.markDirty();
+					}
+				}
+
+			}
+		}
 	}
 	
 	/** Puts a given tile into given map at specified coordinates and,
